Refactored queries where navigational property Problems from Contest is used to use ProblemGroups https://github.com/SoftUni-Internal/suls-issues/issues/3779

+++ Open Judge System/Web/OJS.Web/Areas/Contests/ViewModels/Contests/OnlineContestConfirmViewModel.cs
@@ -10,6 +10,6 @@
 
         public TimeSpan ContestDuration { get; set; }
 
-        public short ProblemGroupsCount { get; set; }
+        public int ProblemGroupsCount { get; set; }
     }
 }
